Change api server project structure
