+++ api/Pipfile
@@ -4,7 +4,6 @@ verify_ssl = true
 name = "pypi"
 
 [packages]
-pytest = "==3.7.3"
 lxml = ">=4.1.1"
 prettyprint = ">=0.1.5"
 "psycopg2" = ">=2.7"
@@ -18,6 +17,8 @@ Flask-SQLAlchemy = ">=2.3"
 Flask-Caching = ">=1.3.3"
 Flask-Cors = ">=3.0"
 Wikidata = ">=0.6.1"
+flask-restplus = "*"
+pip = "==18.0"
 
 [dev-packages]
