+++ api/app/controllers/__init__.py
@@ -1,0 +1,6 @@
+from .base import *
+from .occasion import *
+from .sitemap import *
+from .tags import *
+from .thesis import *
+from .react import *
