+++ api/app/services/__init__.py
@@ -1,0 +1,9 @@
+
+
+
+
+from flask_sqlalchemy import SQLAlchemy
+from flask_caching import Cache
+
+db = SQLAlchemy()
+cache = Cache()
