+++ api/tests/test_occasions.tavern.yaml
@@ -3,7 +3,7 @@ test_name: occasions endpoint
 stages:
   - name: Make sure that occasions list is returned
     request:
-      url: http://localhost:9000/api/v1/occasions/
+      url: http://localhost:9000/api/v2/occasions/
       method: GET
     response:
       status_code: 200
@@ -12,7 +12,7 @@ stages:
 
   - name: Make sure that occasions list including thesis data is returned
     request:
-      url: http://localhost:9000/api/v1/occasions/?thesis_data=1
+      url: http://localhost:9000/api/v2/occasions/?thesis_data=1
       method: GET
     response:
       status_code: 200
@@ -21,7 +21,7 @@ stages:
 
   - name: Make sure that specific occasion is returned
     request:
-      url: http://localhost:9000/api/v1/occasions/1
+      url: http://localhost:9000/api/v2/occasions/1
       method: GET
     response:
       status_code: 200
@@ -30,7 +30,7 @@ stages:
 
   - name: Make sure that unknown occasion gives 404
     request:
-      url: http://localhost:9000/api/v1/occasions/999
+      url: http://localhost:9000/api/v2/occasions/999
       method: GET
     response:
       status_code: 404
