+++ api/tests/test_react.tavern.yaml
@@ -3,7 +3,7 @@ test_name: react endpoint
 stages:
   - name: Make sure that GET requests are disabled
     request:
-      url: http://localhost:9000/api/v1/react/thesis-report
+      url: http://localhost:9000/api/v2/react/thesis-report
       method: GET
     response:
       status_code: 405
@@ -12,7 +12,7 @@ stages:
 
   - name: Make sure that invalid reaction types are not handled
     request:
-      url: http://localhost:9000/api/v1/react/schmu
+      url: http://localhost:9000/api/v2/react/schmu
       method: POST
     response:
       status_code: 200
@@ -23,7 +23,7 @@ stages:
 
   - name: Make sure that thesis reports are handled
     request:
-      url: http://localhost:9000/api/v1/react/thesis-report
+      url: http://localhost:9000/api/v2/react/thesis-report
       method: POST
       json:
         thesis_id: WOM-001-01
@@ -36,7 +36,7 @@ stages:
 
   - name: Make sure that thesis reports missing thesis id return an error
     request:
-      url: http://localhost:9000/api/v1/react/thesis-report
+      url: http://localhost:9000/api/v2/react/thesis-report
       method: POST
       json:
         uuid: test-uuid
@@ -50,7 +50,7 @@ stages:
 
   - name: Make sure that thesis reports missing uuid return an error
     request:
-      url: http://localhost:9000/api/v1/react/thesis-report
+      url: http://localhost:9000/api/v2/react/thesis-report
       method: POST
       json:
         thesis_id: WOM-001-01
@@ -64,7 +64,7 @@ stages:
 
   - name: Make sure that new reactions are handled
     request:
-      url: http://localhost:9000/api/v1/react/reaction
+      url: http://localhost:9000/api/v2/react/reaction
       method: POST
       json:
         thesis_id: WOM-001-01
@@ -80,7 +80,7 @@ stages:
 
   - name: Make sure that duplicate reactions are handled
     request:
-      url: http://localhost:9000/api/v1/react/reaction
+      url: http://localhost:9000/api/v2/react/reaction
       method: POST
       json:
         thesis_id: WOM-001-01
@@ -97,7 +97,7 @@ stages:
 
   - name: Make sure that reactions missing thesis id return an error
     request:
-      url: http://localhost:9000/api/v1/react/reaction
+      url: http://localhost:9000/api/v2/react/reaction
       method: POST
       json:
         uuid: test-uuid
@@ -111,7 +111,7 @@ stages:
 
   - name: Make sure that reactions missing uuid return an error
     request:
-      url: http://localhost:9000/api/v1/react/reaction
+      url: http://localhost:9000/api/v2/react/reaction
       method: POST
       json:
         thesis_id: WOM-001-01
@@ -125,7 +125,7 @@ stages:
 
   - name: Make sure that reactions missing rating return an error
     request:
-      url: http://localhost:9000/api/v1/react/reaction
+      url: http://localhost:9000/api/v2/react/reaction
       method: POST
       json:
         thesis_id: WOM-001-01
